KAN-20 Refactor: Categorize team classes - DrawManagerImpl to engine - RoundState to engine - TwoPlayerMode to screen and rename it to TwoPlayerGameScreen - Background to engine
