+++ src/engine/DrawManagerImpl.java
@@ -1,7 +1,5 @@
-package HUDTeam;
+package engine;
 
-import engine.DrawManager;
-import screen.GameScreen;
 import screen.Screen;
 import entity.Entity;
 import java.awt.Color;
