+++ src/screen/ReceiptScreen.java
@@ -1,6 +1,6 @@
 package screen;
 
-import CtrlS.RoundState;
+import engine.RoundState;
 import engine.GameState;
 
 import java.awt.event.KeyEvent;
